LayoutTests/accessibility/text-role-with-aria-hidden-inside-expected.txt

This tests that when calculating the text under an element, aria-hidden is observed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS text.stringValue is 'AXValue: Song Title Artist Name (4 stars, 1,234 ratings)'
PASS successfullyParsed is true

TEST COMPLETE
